# src/WebApi/spec/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Widget Service",
    "version": "1.0.0"
  },
...
  "paths": {
    "/gadgets": {
      "get": {
        "tags": [
          "Gadgets"
        ],
        "operationId": "GadgetsInterface_getGadgets",
        "parameters": [
          {
            "$ref": "#/components/parameters/APILibrary.ApiVersion"
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "value": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Gadget"
                      },
                      "x-typespec-name": "Gadget[]",
                      "description": "List of elements"
                    },
                    "nextLink": {
                      "type": "string",
                      "format": "uri",
                      "description": "A link to the next page of results if present."
                    }
                  },
                  "required": [
                    "value"
                  ],
                  "x-typespec-name": "APILibrary.List<Gadget>"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "headers": {
              "x-ms-error-code": {
                "required": true,
                "description": "A server-defined code that uniquely identifies the error.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APILibrary.Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Gadgets"
        ],
        "operationId": "GadgetsInterface_createGadget",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gadget"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "headers": {
              "x-ms-error-code": {
                "required": true,
                "description": "A server-defined code that uniquely identifies the error.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APILibrary.Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GadgetCreate"
              }
            }
          }
        }
      }
    },
    "/gadgets/{id}": {
      "get": {
        "tags": [
          "Gadgets"
        ],
        "operationId": "GadgetsInterface_getGadgetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/APILibrary.ApiVersion"
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gadget"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "headers": {
              "x-ms-error-code": {
                "required": true,
                "description": "A server-defined code that uniquely identifies the error.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APILibrary.Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Gadgets"
        ],
        "operationId": "GadgetsInterface_updateGadget",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/APILibrary.ApiVersion"
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Gadget"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "headers": {
              "x-ms-error-code": {
                "required": true,
                "description": "A server-defined code that uniquely identifies the error.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APILibrary.Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/merge-patch+json": {
              "schema": {
                "type": "object",
                "properties": {
                  "maxRPM": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "color": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "Yellow"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "Green"
                        ]
                      }
                    ],
                    "x-typespec-name": "APILibrary.Color.Yellow | APILibrary.Color.Green"
                  },
                  "manufacturer": {
                    "$ref": "#/components/schemas/ManufacturerUpdate"
                  }
                },
                "x-typespec-name": "APILibrary.MergePatch<Gadget>"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Gadgets"
        ],
        "operationId": "GadgetsInterface_deleteGadget",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/APILibrary.ApiVersion"
          }
        ],
        "responses": {
          "204": {
            "description": "There is no content to send for this request, but the headers may be useful. "
          },
          "default": {
            "description": "An unexpected error response.",
            "headers": {
              "x-ms-error-code": {
                "required": true,
                "description": "A server-defined code that uniquely identifies the error.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APILibrary.Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "APILibrary.ApiVersion": {
        "name": "api-version",
        "in": "query",
        "required": true,
        "description": "The version of the API in the form YYYY-MM-DD",
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
      "APILibrary.Color": {
        "type": "string",
        "enum": [
          "Red",
          "Yellow",
          "Purple",
          "Green",
          "Orange",
          "Blue"
        ]
      },
      "APILibrary.Error": {
        "type": "object",
        "properties": {
          "error": {
            "allOf": [
              {
                "$ref": "#/components/schemas/APILibrary.ErrorDetail"
              }
            ],
            "description": "Top-level error object"
          }
        },
        "description": "Error response",
        "required": [
          "error"
        ]
      },
      "APILibrary.ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "A server-defined code that uniquely identifies the error."
          },
          "message": {
            "type": "string",
            "description": "A human-readable representation of the error."
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/APILibrary.ErrorDetail"
            },
            "x-typespec-name": "APILibrary.ErrorDetail[]",
            "description": "An array of details about specific errors that led to this reported error."
          }
        },
        "description": "Error details",
        "required": [
          "code",
          "message"
        ]
      },
      "BaseThingymajig": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "color": {
            "$ref": "#/components/schemas/APILibrary.Color"
          },
          "manufacturer": {
            "$ref": "#/components/schemas/Manufacturer"
          },
          "createdOn": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time that the resource was created",
            "readOnly": true
          },
          "updatedOn": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time that the resource was last updated",
            "readOnly": true
          }
        },
        "required": [
          "id",
          "manufacturer",
          "createdOn",
          "updatedOn"
        ]
      },
      "BaseThingymajigCreate": {
        "type": "object",
        "properties": {
          "color": {
            "$ref": "#/components/schemas/APILibrary.Color"
          },
          "manufacturer": {
            "$ref": "#/components/schemas/Manufacturer"
          }
        },
        "required": [
          "manufacturer"
        ]
      },
...
      "Gadget": {
        "type": "object",
        "properties": {
          "maxRPM": {
            "type": "integer",
            "format": "int32"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "Yellow"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "Green"
                ]
              }
            ],
            "x-typespec-name": "APILibrary.Color.Yellow | APILibrary.Color.Green"
          }
        },
        "required": [
          "maxRPM"
        ],
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseThingymajig"
          }
        ]
      },
      "GadgetCreate": {
        "type": "object",
        "properties": {
          "maxRPM": {
            "type": "integer",
            "format": "int32"
          },
          "color": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "Yellow"
                ]
              },
              {
                "type": "string",
                "enum": [
                  "Green"
                ]
              }
            ],
            "x-typespec-name": "APILibrary.Color.Yellow | APILibrary.Color.Green"
          }
        },
        "required": [
          "maxRPM"
        ],
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseThingymajigCreate"
          }
        ]
      },
...
      "Manufacturer": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "address"
        ]
      },
      "ManufacturerUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        }
      },
...
    }
  }
}